+++ z0bug_odoo/travis/get_test_dependencies.py
@@ -24,7 +24,7 @@ except ImportError:
     from .test_server import get_test_dependencies
 
 
-__version__ = '2.0.13'
+__version__ = '2.0.14'
 
 
 def get_module_list(paths):
@@ -122,5 +122,6 @@ def main(argv=None):
 
 if __name__ == "__main__":
     sys.exit(main())
+