+++ z0bug_odoo/z0bug_odoo_lib.py
@@ -23,7 +23,7 @@ from openpyxl import load_workbook
 
 from python_plus import unicodes
 
-__version__ = "2.0.13"
+__version__ = "2.0.14"
 
 
 class Z0bugOdoo(object):
@@ -40,7 +40,8 @@ class Z0bugOdoo(object):
             except ImportError:
                 self.release = None
         self.model_list = []
-        self.declare_data_dir(None)
+        self.caller_data_dir = self.default_data_dir = None
+        self.declare_data_dir()
 
     def get_image_filename(self, xref):
         file_image = os.path.join(os.path.dirname(__file__), 'data', '%s.png' % xref)
@@ -70,11 +71,10 @@ class Z0bugOdoo(object):
                 row["code"] = "%s" % row["code"]
             getattr(self, pymodel)[row['id']] = unicodes(row)
 
-    def get_data_file_xlsx(self, model, fqn):
+    def get_data_file_xlsx(self, model, fqn, sheet_name=None):
         pymodel = self.get_pymodel(model)
         wb = load_workbook(fqn)
-        for sheet in wb:
-            break
+        sheet = wb.active if not sheet_name else wb[sheet_name]
         colnames = []
         for column in sheet.columns:
             colnames.append(column[0].value)
@@ -133,52 +133,50 @@ class Z0bugOdoo(object):
         fqn = self.get_data_filename(
             filename, bin_types=["xlsx", "csv"], raise_if_not_found=raise_if_not_found)
         self.model_list.append(model)
-        return (self.get_data_file_xlsx(model, fqn)
-                if fqn.endswith(".xlsx")
-                else self.get_data_file_csv(model, fqn))
+        if fqn and fqn.endswith(".xlsx"):
+            self.get_data_file_xlsx(model, fqn)
+        elif fqn and fqn.endswith(".csv"):
+            self.get_data_file_csv(model, fqn)
+        else:
+            setattr(self, self.get_pymodel(model), {})
 
     def get_test_xrefs(self, model):
         """Return model xref list"""
         pymodel = self.get_pymodel(model)
-        if not hasattr(self, pymodel):
-            self.get_data_file(model, pymodel)
-        return list(getattr(self, pymodel))
+        if not hasattr(self, pymodel) or not getattr(self, pymodel):
+            self.get_data_file(model, pymodel, raise_if_not_found=False)
+        return list(getattr(self, pymodel).keys())
 
-    def get_test_values(self, model, xref):
+    def get_test_values(self, model, xref, raise_if_not_found=True):
         """Return model values for test"""
         xids = xref.split('.')
         if len(xids) == 1:
             xids[0], xids[1] = 'z0bug', xids[0]
-        pymodel = self.get_pymodel(model)
-        if not hasattr(self, pymodel):
-            self.get_data_file(model, pymodel)
-        if xref not in getattr(self, pymodel):
-            if xids[0] == 'z0bug':
+        if xref not in self.get_test_xrefs(model):
+            if xids[0] == 'z0bug' and raise_if_not_found:
                 raise KeyError('Invalid xref %s for model %s!' % (xref, model))
             return {}
-        return getattr(self, pymodel)[xref]
+        return getattr(self, self.get_pymodel(model))[xref]
 
     def initialize_model(self, model, merge=False):
         """Write all record of model with test values"""
         pymodel = self.get_pymodel(model)
         if not merge and hasattr(self, pymodel):
             delattr(self, pymodel)
-        # if not hasattr(self, pymodel):
-        #     self.get_data_file(model, pymodel)
 
-    def declare_data_dir(self, data_dir, merge=False, raise_if_not_found=True):
+    def declare_data_dir(self, data_dir=None, merge=False, raise_if_not_found=True):
         if data_dir:
             if not os.path.isdir(data_dir):
                 if raise_if_not_found:
                     raise KeyError("Directory %s not found in the system" % data_dir)
                 return
-            old_caller_data_dir = self.caller_data_dir
-            old_default_data_dir = self.default_data_dir
-        else:
-            old_caller_data_dir = old_default_data_dir = None
-        self.default_data_dir = (os.path.join(os.path.dirname(__file__), 'data')
-                                 if not data_dir or merge else None)
+        old_caller_data_dir = self.caller_data_dir
         self.caller_data_dir = data_dir
+        old_default_data_dir = self.default_data_dir
+        self.default_data_dir = (
+            os.path.join(os.path.dirname(__file__), 'data')
+            if merge or (self.caller_data_dir is None
+                         and self.default_data_dir is None) else None)
         if (
                 self.caller_data_dir != old_caller_data_dir
                 or self.default_data_dir != old_default_data_dir